add employee get by id

+++ src/main/java/com/example/springdemo/repository/EmployeeRepository.java
@@ -3,5 +3,6 @@ package com.example.springdemo.repository;
 import com.example.springdemo.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
